refactor: Phase 2 - collapse getSessionLogData if-chains into adapter pattern
- Replace 6-ecosystem buildTurns if-chain in getSessionLogData() with
  adapter delegation (~450 lines -> ~20 lines)
- Fix ClaudeDesktopAdapter constructor call with required callbacks
- Replace openRawFile VS-specific handler with adapter getRawFileContent()
- Fix detectEditorSource to use findEcosystem() instead of isOpenCodeSessionFile()
