+++ vscode-extension/src/extension.ts
@@ -1,4 +1,4 @@
-import * as vscode from 'vscode';
+﻿import * as vscode from 'vscode';
 import * as fs from 'fs';
 import * as path from 'path';
 import * as os from 'os';
@@ -840,7 +840,7 @@ class CopilotTokenTracker implements vscode.Disposable {
 			new CrushAdapter(this.crush),
 			new VisualStudioAdapter(this.visualStudio, (t, m) => this.estimateTokensFromText(t, m)),
 			new ContinueAdapter(this.continue_),
-			new ClaudeDesktopAdapter(this.claudeDesktopCowork),
+			new ClaudeDesktopAdapter(this.claudeDesktopCowork, (t) => this.isMcpTool(t), (t) => this.extractMcpServerName(t), (t, m) => this.estimateTokensFromText(t, m)),
 			new ClaudeCodeAdapter(this.claudeCode),
 			new MistralVibeAdapter(this.mistralVibe),
 		];
@@ -3227,7 +3227,7 @@ class CopilotTokenTracker implements vscode.Disposable {
 	 * Detect which editor the session file belongs to based on its path.
 	 */
 	private detectEditorSource(filePath: string): string {
-		return _detectEditorSource(filePath, (p) => this.openCode.isOpenCodeSessionFile(p));
+		return _detectEditorSource(filePath, (p) => !!this.findEcosystem(p));
 	}
 
 	/**
@@ -3238,457 +3238,27 @@ class CopilotTokenTracker implements vscode.Disposable {
 		const turns: ChatTurn[] = [];
 
 		try {
-			// Handle OpenCode sessions
-			if (this.openCode.isOpenCodeSessionFile(sessionFile)) {
-				const messages = await this.openCode.getOpenCodeMessagesForSession(sessionFile);
-				if (messages.length > 0) {
-					let turnNumber = 0;
-					let prevCumulativeTotal = 0; // track cumulative total to compute per-turn deltas
-					for (let i = 0; i < messages.length; i++) {
-						const msg = messages[i];
-						if (msg.role !== 'user') { continue; }
-						turnNumber++;
-						// Collect ALL assistant messages for this turn (agentic tool-use loops produce multiple)
-						const turnAssistantMsgs = messages.filter((m, idx) => idx > i && m.role === 'assistant' && m.parentID === msg.id);
-						const userParts = await this.openCode.getOpenCodePartsForMessage(msg.id);
-						const userText = userParts.filter(p => p.type === 'text').map(p => p.text || '').join('\n');
-						let assistantText = '';
-						let thinkingText = '';
-						const toolCalls: { toolName: string; arguments?: string; result?: string }[] = [];
-						let model: string | null = null;
-						let thinkingTokens = 0;
-
-						// Process all assistant messages in this turn to collect text, tool calls, and token totals
-						let turnCumulativeTotal = prevCumulativeTotal;
-						for (const assistantMsg of turnAssistantMsgs) {
-							if (!model) {
-								model = assistantMsg.modelID || null;
-							}
-							thinkingTokens += assistantMsg.tokens?.reasoning || 0;
-							// Track the cumulative total — the last assistant message has the highest value
-							if (typeof assistantMsg.tokens?.total === 'number') {
-								turnCumulativeTotal = Math.max(turnCumulativeTotal, assistantMsg.tokens.total);
-							}
-							const assistantParts = await this.openCode.getOpenCodePartsForMessage(assistantMsg.id);
-							for (const part of assistantParts) {
-								if (part.type === 'text' && part.text) {
-									assistantText += part.text;
-								} else if (part.type === 'reasoning' && part.text) {
-									thinkingText += part.text;
-								} else if (part.type === 'tool' && part.tool) {
-									toolCalls.push({
-										toolName: part.tool,
-										arguments: part.state?.input ? JSON.stringify(part.state.input) : undefined,
-										result: part.state?.output || undefined
-									});
-								}
-							}
-						}
-
-						// Per-turn tokens = delta of cumulative total between this turn and previous
-						const turnTokens = turnCumulativeTotal - prevCumulativeTotal;
-						// Split proportionally: output+thinking are known, remainder is input
-						const turnOutputAndThinking = turnAssistantMsgs.reduce((sum, m) => sum + (m.tokens?.output || 0) + (m.tokens?.reasoning || 0), 0);
-						const turnInputTokens = Math.max(0, turnTokens - turnOutputAndThinking);
-
-						turns.push({
-							turnNumber,
-							timestamp: msg.time?.created ? new Date(msg.time.created).toISOString() : null,
-							mode: (msg.agent === 'build' || msg.agent === 'agent') ? 'agent' : (msg.agent === 'ask' ? 'ask' : 'agent'),
-							userMessage: userText,
-							assistantResponse: assistantText,
-							model,
-							toolCalls,
-							contextReferences: {
-								file: 0, selection: 0, implicitSelection: 0, symbol: 0, codebase: 0,
-								workspace: 0, terminal: 0, vscode: 0,
-								terminalLastCommand: 0, terminalSelection: 0, clipboard: 0, changes: 0,
-								outputPanel: 0, problemsPanel: 0, byKind: {}, copilotInstructions: 0, agentsMd: 0, byPath: {}
-							},
-							mcpTools: [],
-							inputTokensEstimate: turnInputTokens,
-							outputTokensEstimate: turnOutputAndThinking - thinkingTokens,
-							thinkingTokensEstimate: thinkingTokens
-						});
-
-						prevCumulativeTotal = turnCumulativeTotal;
-					}
-				}
-				return {
-					file: details.file,
-					title: details.title || null,
-					editorSource: details.editorSource,
-					editorName: details.editorName || 'OpenCode',
-					size: details.size,
-					modified: details.modified,
-					interactions: details.interactions,
-					contextReferences: details.contextReferences,
-					firstInteraction: details.firstInteraction,
-					lastInteraction: details.lastInteraction,
-					turns,
-					usageAnalysis: undefined
-				};
-			}
-
-		// Handle Visual Studio sessions
-		if (this.visualStudio.isVSSessionFile(sessionFile)) {
-			const objects = this.visualStudio.decodeSessionFile(sessionFile);
-			let turnNumber = 0;
-			for (let i = 1; i < objects.length; i += 2) {
-				const req = objects[i];
-				const res = objects[i + 1];
-				if (!req) { continue; }
-				const reqData = req[1];
-				const resData = res?.[1];
-				turnNumber++;
-				const userText = this.visualStudio.extractTextFromContent(reqData?.Content || []);
-				const assistantText = res ? this.visualStudio.extractTextFromContent(resData?.Content || []) : '';
-				const model = this.visualStudio.getModelId(resData ?? reqData, !resData);
-				const contextText = this.visualStudio.extractContextText(reqData?.Context);
-								const inputTokens = this.estimateTokensFromText(userText + contextText, model ?? 'gpt-4');
-				const outputTokens = res ? this.estimateTokensFromText(assistantText, model ?? 'gpt-4') : 0;
-				const toolCalls: { toolName: string; arguments?: string; result?: string }[] = [];
-				for (const c of (resData?.Content || [])) {
-					const inner = Array.isArray(c) ? c[1] : null;
-					if (inner?.Function) {
-						toolCalls.push({
-							toolName: String(inner.Function.Description || 'tool'),
-							result: typeof inner.Function.Result === 'string' ? inner.Function.Result : undefined
-						});
-					}
-				}
-				turns.push({
-					turnNumber,
-					timestamp: reqData?.Timestamp ? new Date(reqData.Timestamp).toISOString() : null,
-					mode: 'ask' as const,
-					userMessage: userText,
-					assistantResponse: assistantText,
-					model,
-					toolCalls,
-					contextReferences: {
-						file: 0, selection: 0, implicitSelection: 0, symbol: 0, codebase: 0,
-						workspace: 0, terminal: 0, vscode: 0,
-						terminalLastCommand: 0, terminalSelection: 0, clipboard: 0, changes: 0,
-						outputPanel: 0, problemsPanel: 0, byKind: {}, copilotInstructions: 0, agentsMd: 0, byPath: {}
-					},
-					mcpTools: [],
-					inputTokensEstimate: inputTokens,
-					outputTokensEstimate: outputTokens,
-					thinkingTokensEstimate: 0
-				});
-			}
-			return {
-				file: details.file,
-				title: details.title || null,
-				editorSource: details.editorSource,
-				editorName: 'Visual Studio',
-				size: details.size,
-				modified: details.modified,
-				interactions: details.interactions,
-				contextReferences: details.contextReferences,
-				firstInteraction: details.firstInteraction,
-				lastInteraction: details.lastInteraction,
-				turns,
-				usageAnalysis: undefined
-			};
-		}
-
-// Handle Crush sessions
-			if (this.crush.isCrushSessionFile(sessionFile)) {
-				const messages = await this.crush.getCrushMessages(sessionFile);
-				const session = await this.crush.readCrushSession(sessionFile);
-				const totalTokens = (session?.prompt_tokens ?? 0) + (session?.completion_tokens ?? 0);
-				const userMessages = messages.filter(m => m.role === 'user');
-				const numTurns = userMessages.length;
-				let turnNumber = 0;
-				for (let i = 0; i < messages.length; i++) {
-					const msg = messages[i];
-					if (msg.role !== 'user') { continue; }
-					turnNumber++;
-					// Collect assistant messages until the next user message
-					const turnAssistantMsgs: any[] = [];
-					for (let j = i + 1; j < messages.length; j++) {
-						if (messages[j].role === 'user') { break; }
-						if (messages[j].role === 'assistant') { turnAssistantMsgs.push(messages[j]); }
-					}
-					// Extract user text from parts
-					const userParts: any[] = Array.isArray(msg.parts) ? msg.parts : [];
-					const userText = userParts
-						.filter(p => p?.type === 'text' && p?.text)
-						.map(p => p.text as string)
-						.join('\n');
-					let assistantText = '';
-					const toolCalls: { toolName: string; arguments?: string; result?: string }[] = [];
-					let model: string | null = null;
-					for (const assistantMsg of turnAssistantMsgs) {
-						if (!model) { model = assistantMsg.model || null; }
-						const parts: any[] = Array.isArray(assistantMsg.parts) ? assistantMsg.parts : [];
-						for (const part of parts) {
-							if (part?.type === 'text' && part?.text) {
-								assistantText += part.text;
-							} else if (part?.type === 'tool_call' && part?.data?.name) {
-								toolCalls.push({
-									toolName: part.data.name,
-									arguments: part.data.arguments ? JSON.stringify(part.data.arguments) : undefined
-								});
-							}
-						}
-					}
-					// Distribute session token totals evenly across turns
-					const perTurnTokens = numTurns > 0 ? Math.round(totalTokens / numTurns) : 0;
-					const perTurnInput = session?.prompt_tokens && numTurns > 0 ? Math.round(session.prompt_tokens / numTurns) : 0;
-					const perTurnOutput = session?.completion_tokens && numTurns > 0 ? Math.round(session.completion_tokens / numTurns) : 0;
-					turns.push({
-						turnNumber,
-						timestamp: msg.created_at ? new Date(msg.created_at * 1000).toISOString() : null,
-						mode: 'agent',
-						userMessage: userText,
-						assistantResponse: assistantText,
-						model,
-						toolCalls,
-						contextReferences: {
-							file: 0, selection: 0, implicitSelection: 0, symbol: 0, codebase: 0,
-							workspace: 0, terminal: 0, vscode: 0,
-							terminalLastCommand: 0, terminalSelection: 0, clipboard: 0, changes: 0,
-							outputPanel: 0, problemsPanel: 0, byKind: {}, copilotInstructions: 0, agentsMd: 0, byPath: {}
-						},
-						mcpTools: [],
-						inputTokensEstimate: perTurnInput,
-						outputTokensEstimate: perTurnOutput,
-						thinkingTokensEstimate: 0
-					});
-				}
-				return {
-					file: details.file,
-					title: details.title || null,
-					editorSource: details.editorSource,
-					editorName: 'Crush',
-					size: details.size,
-					modified: details.modified,
-					interactions: details.interactions,
-					contextReferences: details.contextReferences,
-					firstInteraction: details.firstInteraction,
-					lastInteraction: details.lastInteraction,
-					turns,
-					usageAnalysis: undefined
-				};
-			}
-
-			// Handle Continue sessions
-			if (this.continue_.isContinueSessionFile(sessionFile)) {
-				const continueTurns = this.continue_.buildContinueTurns(sessionFile);
-				const emptyContextRefs = _createEmptyContextRefs();
-				for (const ct of continueTurns) {
-					turns.push({
-						turnNumber: turns.length + 1,
-						timestamp: null,
-						mode: 'ask',
-						userMessage: ct.userText,
-						assistantResponse: ct.assistantText,
-						model: ct.model,
-						toolCalls: ct.toolCalls,
-						contextReferences: emptyContextRefs,
-						mcpTools: [],
-						inputTokensEstimate: ct.inputTokens,
-						outputTokensEstimate: ct.outputTokens,
-						thinkingTokensEstimate: 0
-					});
-				}
-				return {
-					file: details.file,
-					title: details.title || null,
-					editorSource: details.editorSource,
-					editorName: details.editorName || 'Continue',
-					size: details.size,
-					modified: details.modified,
-					interactions: details.interactions,
-					contextReferences: details.contextReferences,
-					firstInteraction: details.firstInteraction,
-					lastInteraction: details.lastInteraction,
-					turns,
-					usageAnalysis: undefined
-				};
-			}
-
-			// Handle Claude Desktop Cowork sessions
-			if (this.claudeDesktopCowork.isCoworkSessionFile(sessionFile)) {
-				const events = this.claudeDesktopCowork.readCoworkEvents(sessionFile);
-				let currentUserEvent: any = null;
-				const pendingAssistantEvents: any[] = [];
-
-				const emitTurn = () => {
-					if (!currentUserEvent) { return; }
-					const content = currentUserEvent.message?.content;
-					const userMessage = typeof content === 'string' ? content
-						: Array.isArray(content) ? content.filter((c: any) => c.type === 'text').map((c: any) => c.text || '').join('\n')
-						: '';
-					let assistantText = '';
-					let actualInputTokens = 0;
-					let actualOutputTokens = 0;
-					let model: string | null = null;
-					const toolCalls: { toolName: string; arguments?: string }[] = [];
-					const mcpTools: { server: string; tool: string }[] = [];
-
-					for (const ae of pendingAssistantEvents) {
-						const msg = ae.message;
-						if (!model && msg?.model) { model = msg.model; }
-						const usage = msg?.usage;
-						if (usage) {
-							actualInputTokens += (usage.input_tokens || 0) + (usage.cache_creation_input_tokens || 0) + (usage.cache_read_input_tokens || 0);
-							actualOutputTokens += usage.output_tokens || 0;
-						}
-						const contentArr: any[] = Array.isArray(msg?.content) ? msg.content : [];
-						for (const block of contentArr) {
-							if (block.type === 'text') { assistantText += block.text || ''; }
-							else if (block.type === 'tool_use') {
-								const toolName: string = block.name || 'unknown';
-								if (this.isMcpTool(toolName)) {
-									mcpTools.push({ server: this.extractMcpServerName(toolName), tool: toolName });
-								} else {
-									toolCalls.push({ toolName, arguments: block.input ? JSON.stringify(block.input) : undefined });
-								}
-							}
-						}
-					}
-
-					const usedModel = model || 'claude-sonnet-4-6';
-					const actualUsage: ActualUsage | undefined = (actualInputTokens > 0 || actualOutputTokens > 0) ? {
-						promptTokens: actualInputTokens,
-						completionTokens: actualOutputTokens
-					} : undefined;
-
-					turns.push({
-						turnNumber: turns.length + 1,
-						timestamp: currentUserEvent.timestamp ? new Date(currentUserEvent.timestamp).toISOString() : null,
-						mode: 'agent',
-						userMessage,
-						assistantResponse: assistantText,
-						model: usedModel,
-						toolCalls,
-						contextReferences: _createEmptyContextRefs(),
-						mcpTools,
-						inputTokensEstimate: actualInputTokens || this.estimateTokensFromText(userMessage, usedModel),
-						outputTokensEstimate: actualOutputTokens || this.estimateTokensFromText(assistantText, usedModel),
-						thinkingTokensEstimate: 0,
-						actualUsage
-					});
-				};
-
-				const isRealUserMessage = (event: any): boolean => {
-					const content = event.message?.content;
-					if (typeof content === 'string') { return !!content.trim(); }
-					if (!Array.isArray(content)) { return false; }
-					const hasText = content.some((c: any) => c.type === 'text');
-					const hasToolResult = content.some((c: any) => c.type === 'tool_result');
-					return hasText && !hasToolResult;
-				};
-
-				for (const event of events) {
-					if (event.type === 'user' && !event.isSidechain && event.message?.role === 'user' && isRealUserMessage(event)) {
-						emitTurn();
-						currentUserEvent = event;
-						pendingAssistantEvents.length = 0;
-					} else if (event.type === 'assistant' && event.message?.stop_reason && event.message?.role === 'assistant') {
-						// Only collect final (non-streaming) assistant events — stop_reason is '' on fragments
-						pendingAssistantEvents.push(event);
-					}
-				}
-				emitTurn();
-
-				return {
-					file: details.file,
-					title: details.title || null,
-					editorSource: details.editorSource,
-					editorName: 'Claude Desktop Cowork',
-					size: details.size,
-					modified: details.modified,
-					interactions: details.interactions,
-					contextReferences: details.contextReferences,
-					firstInteraction: details.firstInteraction,
-					lastInteraction: details.lastInteraction,
-					turns,
-					usageAnalysis: undefined
-				};
-			}
-
-			// Handle Mistral Vibe sessions
-			if (this.mistralVibe.isVibeSessionFile(sessionFile)) {
-				const messages = this.mistralVibe.readSessionMessages(sessionFile);
-				const sessionMeta = this.mistralVibe.getSessionMeta(sessionFile);
-				const tokenData = this.mistralVibe.getTokensFromSession(sessionFile);
-				const model: string = sessionMeta.model || 'devstral';
-
-				// Collect non-injected user messages as turn boundaries
-				const userMsgIndices: number[] = [];
-				for (let i = 0; i < messages.length; i++) {
-					if (messages[i].role === 'user' && messages[i].injected !== true) {
-						userMsgIndices.push(i);
-					}
-				}
-				// Mistral Vibe only has session-level token counts (not per-turn)
-				// Set per-turn estimates to 0; actual totals go in actualTokens
-
-				for (let t = 0; t < userMsgIndices.length; t++) {
-					const userIdx = userMsgIndices[t];
-					const nextUserIdx = t + 1 < userMsgIndices.length ? userMsgIndices[t + 1] : messages.length;
-					const userMsg = messages[userIdx];
-					const userText = typeof userMsg.content === 'string' ? userMsg.content : '';
-					let assistantText = '';
-					const toolCalls: { toolName: string; arguments?: string; result?: string }[] = [];
-
-					// Collect all messages between this user message and the next non-injected user message
-					for (let j = userIdx + 1; j < nextUserIdx; j++) {
-						const msg = messages[j];
-						if (msg.role === 'assistant') {
-							if (typeof msg.content === 'string') { assistantText += msg.content; }
-							if (Array.isArray(msg.tool_calls)) {
-								for (const tc of msg.tool_calls) {
-									toolCalls.push({
-										toolName: tc.function?.name || tc.name || 'unknown',
-										arguments: tc.function?.arguments ? JSON.stringify(tc.function.arguments) : undefined
-									});
-								}
-							}
-						} else if (msg.role === 'tool') {
-							// Attach tool result to last tool call
-							const last = toolCalls[toolCalls.length - 1];
-							if (last) { last.result = typeof msg.content === 'string' ? msg.content : undefined; }
-						}
-					}
-
-					turns.push({
-						turnNumber: t + 1,
-						timestamp: sessionMeta.firstInteraction,
-						mode: 'agent',
-						userMessage: userText,
-						assistantResponse: assistantText,
-						model,
-						toolCalls,
-						contextReferences: _createEmptyContextRefs(),
-						mcpTools: [],
-						inputTokensEstimate: 0,
-						outputTokensEstimate: 0,
-						thinkingTokensEstimate: 0
-					});
-				}
-
-				return {
-					file: details.file,
-					title: details.title || null,
-					editorSource: details.editorSource,
-					editorName: 'Mistral Vibe',
-					size: details.size,
-					modified: details.modified,
-					interactions: details.interactions,
-					contextReferences: details.contextReferences,
-					firstInteraction: details.firstInteraction,
-					lastInteraction: details.lastInteraction,
-					turns,
-					actualTokens: tokenData.tokens,
-					usageAnalysis: undefined
-				};
-			}
-
+// Delegate to ecosystem adapter if available
+const eco = this.findEcosystem(sessionFile);
+if (eco?.buildTurns) {
+const result = await eco.buildTurns(sessionFile);
+turns.push(...result.turns);
+return {
+file: details.file,
+title: details.title || null,
+editorSource: details.editorSource,
+editorName: details.editorName || eco.displayName,
+size: details.size,
+modified: details.modified,
+interactions: details.interactions,
+contextReferences: details.contextReferences,
+firstInteraction: details.firstInteraction,
+lastInteraction: details.lastInteraction,
+turns,
+...(result.actualTokens !== undefined ? { actualTokens: result.actualTokens } : {}),
+usageAnalysis: undefined
+};
+}
 			const fileContent = await fs.promises.readFile(sessionFile, 'utf8');
 
 			// Check if this is a UUID-only file (new Copilot CLI format)
@@ -4895,12 +4465,10 @@ Return ONLY the JSON object, no markdown formatting, no explanations.`;
 					case 'openRawFile':
 						await this.dispatch('openRawFile:logviewer', async () => {
 							try {
-								if (this.visualStudio.isVSSessionFile(sessionFilePath)) {
-									// VS session files are binary MessagePack — show decoded JSON instead
-									const objects = this.visualStudio.decodeSessionFile(sessionFilePath);
-									const readable = objects.map((obj: any, i: number) => i === 0 ? obj : obj?.[1] ?? obj);
-									const jsonText = JSON.stringify(readable, null, 2);
-									const doc = await vscode.workspace.openTextDocument({ content: jsonText, language: 'json' });
+								const rawEco = this.findEcosystem(sessionFilePath);
+								const rawContent = rawEco?.getRawFileContent?.(sessionFilePath);
+								if (rawContent !== undefined) {
+									const doc = await vscode.workspace.openTextDocument({ content: rawContent, language: 'json' });
 									await vscode.window.showTextDocument(doc);
 								} else {
 									await vscode.window.showTextDocument(vscode.Uri.file(sessionFilePath));
